move obfuscation to global validation pipe
