+++ apps/server/src/core/validation/pipe/global-validation.pipe.ts
@@ -1,5 +1,6 @@
 import { ValidationError, ValidationPipe } from '@nestjs/common';
 import { ApiValidationError } from '@shared/common';
+import { getMetadataStorage } from 'class-validator';
 
 /** *********************************************
  * Global Pipe setup
@@ -10,6 +11,8 @@ import { ApiValidationError } from '@shared/common';
  * validation.
  */
 export class GlobalValidationPipe extends ValidationPipe {
+	private readonly metadataStorage = getMetadataStorage();
+
 	constructor() {
 		super({
 			// enable DTO instance creation for incoming data
@@ -21,11 +24,33 @@ export class GlobalValidationPipe extends ValidationPipe {
 			whitelist: true, // only pass valid @ApiProperty-decorated DTO properties, remove others
 			forbidNonWhitelisted: false, // additional params are just skipped (required when extracting multiple DTO from single query)
 			forbidUnknownValues: true,
-			exceptionFactory: (errors: ValidationError[]) => new ApiValidationError(errors),
+			exceptionFactory: (errors: ValidationError[]) =>
+				new ApiValidationError(this.obfuscatePrivacyProtectedValues(errors)),
 			validationError: {
+				// make sure target (DTO) is set on validation error
+				// we need this to be able to get DTO metadata for the obfuscation check
+				// see obfuscatePrivacyProtectedValues()
 				target: true,
 				value: true,
 			},
+		});
+	}
+
+	private obfuscatePrivacyProtectedValues(errors: ValidationError[]): ValidationError[] {
+		return errors.map((e) => {
+			const metadatas = this.metadataStorage.getTargetValidationMetadatas(e.target!.constructor, '', true, true);
+
+			const obfuscateValue = metadatas.some(
+				(validationMetadata) =>
+					// eslint-disable-next-line @typescript-eslint/no-unsafe-member-access
+					validationMetadata.propertyName === e.property && validationMetadata.context?.privacyProtect
+			);
+
+			if (obfuscateValue) {
+				e.value = '####';
+			}
+
+			return e;
 		});
 	}
 }
