+++ apps/server/src/shared/common/error/api-validation.error.ts
@@ -1,13 +1,8 @@
 import { HttpStatus, ValidationError } from '@nestjs/common';
-import { getMetadataStorage } from 'class-validator';
 import { BusinessError } from './business.error';
 
 export class ApiValidationError extends BusinessError {
-	readonly validationErrors: ValidationError[];
-
-	private readonly metadataStorage = getMetadataStorage();
-
-	constructor(validationErrors: ValidationError[] = []) {
+	constructor(readonly validationErrors: ValidationError[] = []) {
 		super(
 			{
 				type: 'API_VALIDATION_ERROR',
@@ -16,21 +11,5 @@ export class ApiValidationError extends BusinessError {
 			},
 			HttpStatus.BAD_REQUEST
 		);
-
-		this.validationErrors = validationErrors.map((e) => {
-			const metadatas = this.metadataStorage.getTargetValidationMetadatas(e.target!.constructor, '', true, true);
-
-			const obfuscateValue = metadatas.some(
-				(validationMetadata) =>
-					// eslint-disable-next-line @typescript-eslint/no-unsafe-member-access
-					validationMetadata.propertyName === e.property && validationMetadata.context?.privacyProtect
-			);
-
-			if (obfuscateValue) {
-				e.value = '####';
-			}
-
-			return e;
-		});
 	}
 }
